requirements.txt
discord.py
apscheduler
requests
python-dotenv
flask